requirements/base.txt
enforce==0.3.4
python-dateutil==2.8.0
structlog==19.1.0
